refactor: room-event와 item 파일 이동
