+++ src/conversation-events/events-handler/events-handler.module.ts
@@ -4,7 +4,7 @@ import { UsersModule } from 'src/users/users.module';
 import { PeerJsGateway } from './peer-js.gateway';
 import { RoomGateway } from './room.gateway';
 import { ChattingHandlerGateway } from './chatting-handler.gateway';
-import { RoomEventModule } from 'src/rooms/item/room-event/room-event.module';
+import { RoomEventModule } from 'src/rooms/room-event/room-event.module';
 import { RoomSocketModule } from '../room-socket/room-socket.module';
 
 @Module({
