+++ src/rooms/rooms.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { RoomsService } from './rooms.service';
 import { RoomsController } from './rooms.controller';
-import { RoomEventModule } from './item/room-event/room-event.module';
+import { RoomEventModule } from './room-event/room-event.module';
 import { UsersModule } from 'src/users/users.module';
 
 @Module({
